add if for existingRegularApproverRules.length

+++ lib/modules/platform/gitlab/index.ts
@@ -547,11 +547,13 @@ async function ignoreApprovals(pr: number): Promise<void> {
       ({ rule_type, name }) => rule_type !== 'any_approver' && name !== ruleName
     );
 
-    await p.all(
-      existingRegularApproverRules.map((rule) => async (): Promise<void> => {
-        await gitlabApi.deleteJson(`${url}/${rule.id}`);
-      })
-    );
+    if (existingRegularApproverRules.length) {
+      await p.all(
+        existingRegularApproverRules.map((rule) => async (): Promise<void> => {
+          await gitlabApi.deleteJson(`${url}/${rule.id}`);
+        })
+      );
+    }
 
     if (existingAnyApproverRule) {
       await gitlabApi.putJson(`${url}/${existingAnyApproverRule.id}`, {
